+++ packages/hub/src/scope.ts
@@ -438,14 +438,15 @@ export class Scope implements ScopeInterface {
    * This will be called on every set call.
    */
   protected _notifyScopeListeners(): void {
+    // We need this check for this._notifyingListeners to be able to work on scope during updates
+    // If this check is not here we'll produce endless recursion when something is done with the scope
+    // during the callback.
     if (!this._notifyingListeners) {
       this._notifyingListeners = true;
-      setTimeout(() => {
-        this._scopeListeners.forEach(callback => {
-          callback(this);
-        });
-        this._notifyingListeners = false;
+      this._scopeListeners.forEach(callback => {
+        callback(this);
       });
+      this._notifyingListeners = false;
     }
   }